Ilumina_backend/requirements.txt
djangorestframework
django-cors-headers
pandas
openpyxl